initial iter
