+++ llama-index-integrations/memory/llama-index-memory-mem0/llama_index/memory/mem0/__init__.py
@@ -1,3 +1,4 @@
-from llama_index.memory.mem0.base import Mem0
+from llama_index.memory.mem0.chat_engine_memory import Mem0ChatEngineMemory
+from llama_index.memory.mem0.functional_agent_memory import Mem0FunctionalAgentMemory
 
-__all__ = ["Mem0"]
+__all__ = ["Mem0ChatEngineMemory", "Mem0FunctionalAgentMemory"]
